requirements.txt
APScheduler==3.11.0
beautifulsoup4==4.13.4
fastapi==0.115.12
feedparser==6.0.11
google-genai==1.15.0
hdbscan~=0.8.40
html2text~=2025.4.15
lxml==5.4.0
markdown==3.8.0
newspaper3k~=0.2.8
numpy~=2.0.2
openai~=1.78.0
psycopg2-binary==2.9.10
pydantic==2.11.4
python-dotenv==1.1.0
readability-lxml~=0.8.4.1
resend~=2.10.0
scikit-learn~=1.5.2
sentence-transformers==2.6.1
transformers==4.36.2
starlette==0.46.2
toml==0.10.2
uvicorn~=0.34.2
